refactor: split skills into per-hero files
Split the monolithic skills/index.ts (~774 lines) into per-hero files
to prevent merge conflicts when parallel agents implement different
skills.

New structure:
- types.ts: Shared types (SkillDefinition, SkillUsageType, HeroId)
- arythea.ts, tovak.ts, goldyx.ts, norowas.ts, wolfhawk.ts, krang.ts,
  braevalar.ts: Individual hero skill definitions
- index.ts: Re-exports all constants and combines into SKILLS/HERO_SKILLS

External API is fully preserved - no changes needed to importing files.
